LPS-100146 Renames files to .js since we no longer have legacy code in this module
