+++ modules/apps/site-navigation/site-navigation-admin-web/src/main/resources/META-INF/resources/js/ManagementToolbarDefaultEventHandler.js
